# docs/04-api/openapi/frameflow-v1.yaml
openapi: 3.0.3
info:
  title: FrameFlow Select API
  version: 1.0.0
  description: FrameFlow Select 批量质检与 Top-K 优选平台。身份端点沿用 M01（/api/v1/auth、/api/v1/teams）；产品端点为手写 OpenAPI 权威契约，运行时以 springdoc 输出并经验证对齐。
...
paths:
  /api/v1/teams:
    get:
      tags:
      - teams
      summary: 发现当前用户加入的有效团队
      description: 仅返回当前登录用户 status=ACTIVE 的团队成员关系，用于登录后的团队选择；不会返回其他用户或已移除的成员关系。
      operationId: listMyTeams
      responses:
        '200':
          description: 当前用户的团队成员关系列表
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamList'
        '401':
          description: 未认证或 Token 无效/过期
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - bearerAuth: []
    post:
      tags:
      - teams
      summary: 创建团队（创建者自动成为 OWNER；需要 Idempotency-Key）
      description: 任一 ACTIVE 已认证用户均可创建；团队与创建者 OWNER 关系在同一事务写入。
      operationId: createTeam
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          format: uuid
          maxLength: 128
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTeamRequest'
        required: true
      responses:
        '201':
          description: 团队创建成功
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '400':
          description: 参数校验失败或缺少 Idempotency-Key
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证或 Token 无效/过期
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: IDEMPOTENCY_CONFLICT / IDEMPOTENCY_IN_PROGRESS
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - bearerAuth: []
  /api/v1/teams/{teamId}/members:
    get:
      tags:
      - teams
      summary: 团队 ACTIVE 成员列表（仅团队成员）
      operationId: listTeamMembers
      parameters:
      - name: teamId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: 成员列表
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberList'
        '401':
          description: 未认证或 Token 无效/过期
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 团队不存在或无权访问
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - bearerAuth: []
    post:
      tags:
      - teams
      summary: 直接添加已注册成员为 ACTIVE（仅 OWNER；需要 Idempotency-Key）
      description: 已是 ACTIVE 成员返回 409 TEAM_MEMBER_ALREADY_EXISTS；REMOVED 关系则复用原记录并恢复为 ACTIVE。M01 不创建 INVITED 记录。
      operationId: addTeamMember
      parameters:
      - name: teamId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          format: uuid
          maxLength: 128
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMemberRequest'
        required: true
      responses:
        '201':
          description: 成员已添加
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamMember'
        '400':
          description: 参数校验失败或缺少 Idempotency-Key
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证或 Token 无效/过期
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 已认证但角色不足（非 OWNER）
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 团队或目标用户不存在/无权访问
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: TEAM_MEMBER_ALREADY_EXISTS / 幂等冲突
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - bearerAuth: []
...
  /api/v1/teams/{teamId}/members/{userId}:
    delete:
      tags:
      - teams
      summary: 移除成员（仅 OWNER；不能移除自己或最后一名 OWNER）
      operationId: removeTeamMember
      parameters:
      - name: teamId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '204':
          description: 成员已移除
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
        '401':
          description: 未认证或 Token 无效/过期
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 已认证但角色不足（非 OWNER）
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 团队或成员不存在/无权访问
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: TEAM_SELF_REMOVAL_FORBIDDEN / TEAM_LAST_OWNER_CONFLICT
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - bearerAuth: []
    patch:
      tags:
      - teams
      summary: 修改成员角色（仅 OWNER；不能把最后一名 OWNER 降级）
      operationId: updateMemberRole
      parameters:
      - name: teamId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMemberRoleRequest'
        required: true
      responses:
        '200':
          description: 角色已更新
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamMember'
        '400':
          description: 参数校验失败
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证或 Token 无效/过期
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 已认证但角色不足（非 OWNER）
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 团队或成员不存在/无权访问
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: TEAM_LAST_OWNER_CONFLICT 等
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - bearerAuth: []
  /api/v1/teams/{teamId}:
    get:
      tags:
      - teams
      summary: 查看团队（仅团队成员；非成员统一 404）
      operationId: getTeam
      parameters:
      - name: teamId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: 团队信息
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '401':
          description: 未认证或 Token 无效/过期
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 团队不存在或无权访问
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - bearerAuth: []
...
  /api/v1/auth/me:
    get:
      tags:
      - auth
      summary: 当前登录用户
      operationId: currentUser
      responses:
        '200':
          description: 当前用户信息
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: 未认证或 Token 无效/过期
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - bearerAuth: []
...
components:
  schemas:
    CreateTeamRequest:
      type: object
      properties:
        name:
          type: string
          description: 团队名称
          maxLength: 100
          minLength: 1
      required:
      - name
    Team:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
      - createdAt
      - id
      - name
    Error:
      type: object
      properties:
        code:
          type: string
          description: 稳定错误码
          example: AUTH_REQUIRED
        message:
          type: string
          description: 人类可读的错误说明
        requestId:
          type: string
          format: uuid
          description: 必须与响应 X-Request-Id header 完全相同
        traceId:
          type: string
          description: 可观测环境可选
      required:
      - code
      - message
      - requestId
    AddMemberRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          description: 已注册目标用户邮箱；M01 直接创建或恢复 ACTIVE 成员关系，不发送邀请
        role:
          type: string
          description: 初始团队角色
          enum:
          - OWNER
          - OPERATOR
          - REVIEWER
          - VIEWER
      required:
      - email
      - role
    TeamMember:
      type: object
      properties:
        id:
          type: integer
          format: int64
        teamId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        email:
          type: string
          format: email
        role:
          type: string
          description: 团队角色（CLIENT 是项目级角色，M02 起，不在团队角色枚举内）
          enum:
          - OWNER
          - OPERATOR
          - REVIEWER
          - VIEWER
        status:
          type: string
          enum:
          - ACTIVE
      required:
      - email
      - id
      - role
      - status
      - teamId
      - userId
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        displayName:
          type: string
        status:
          type: string
          enum:
          - ACTIVE
          - DISABLED
          - PENDING
        lastLoginAt:
          type: string
          format: date-time
      required:
      - displayName
      - email
      - id
      - status
...
    UpdateMemberRoleRequest:
      type: object
      properties:
        role:
          type: string
          description: 新团队角色
          enum:
          - OWNER
          - OPERATOR
          - REVIEWER
          - VIEWER
      required:
      - role
...
    TeamList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TeamMembershipSummary'
      required:
      - items
    TeamMembershipSummary:
      type: object
      properties:
        teamId:
          type: integer
          format: int64
        name:
          type: string
        role:
          type: string
          description: 团队角色（CLIENT 是项目级角色，M02 起，不在团队角色枚举内）
          enum:
          - OWNER
          - OPERATOR
          - REVIEWER
          - VIEWER
        status:
          type: string
          enum:
          - ACTIVE
        createdAt:
          type: string
          format: date-time
          description: 团队创建时间
      required:
      - createdAt
      - name
      - role
      - status
      - teamId
    MemberList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TeamMember'
      required:
      - items
...
  headers:
    RequestId:
      description: 服务器为本次 HTTP 请求生成的请求 ID
      required: true
      schema:
        type: string
        format: uuid
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT